Track inputs
`Current.t` terms are now functions that are evaluated in some context.
The context collects a list of the inputs used so that they can be
monitored.

This requires a bit of caching, so that nodes that are reachable by two
paths through the dependency graph don't get evaluated twice.