Allow specifying a validity period for build results
After the period expires, an entry will be built again (if still
needed). This is useful to control how often to run `docker pull` for
example, or to rebuild something weekly.

Also, added tests for the cache. Note: now that we use a database file,
the tests cannot be run in parallel. Also, the output is less confusing
this way, as it only says the tests passed when they all did.