Add support for remote Docker engines
If `$DOCKER_HOST` is set, use that as a database key to remember where
the results are. Just because we built on one engine doesn't mean the
result will be available at another.

Also, allow creating multiple instances of the docker plugin, to allow
pipelines that use multiple Docker hosts.