Combine build and publish caches
Before, we had separate caches (and database tables) for builds and for
outputs. Builds take a key and produce a value, whereas outputs take a
key and a value as input.

However, it turned out that some outputs also need to return an
"outcome", meaning that the build cache can be considered as a special
case of the publish cache where the "value" is always `()`.

This commit removes the old build cache implementation and replaces it
with a wrapper around the publish cache (without changing the existing
API).

It also extends the output cache with some features that were only
present in the build cache, but which were wanted anyway:

- Build numbers (so we can track rebuilds).
- Cancelling on user request.
- Auto-cancel when no longer needed.
- Schedules (e.g. republish unchanged outputs regularly).
- Recording of failures on disk.
- Ready and running times.
- Database queries.

Another useful side-effect of this is that the Query page in the web UI
now shows both build and output jobs. Before, only build jobs were
shown.