Hide transitive dependencies
If we have `A -> B -> C` and `A -> C`, then hide `A -> C` because the
dependency on `A` is implied by the dependency on `B` and it clutters
the diagrams having both.

Also, track the inputs to map. This is needed to remove transitive
dependencies, and also avoids the need for a `map` label, which dot
tends to place anywhere it likes.

We now render edges last. Otherwise, dot pulls things into clusters that
shouldn't be there.