Extract useful error messages from logs
OCurrent can now be configured with a list of log patterns and some
replacement text to report if the pattern is found. e.g.

  Pattern: `ERROR: unsatisfiable constraints:.*[\n].* ([^ ]+) \(missing\)`
  Report: `Missing Alpine package \1`

After each build, it finds the highest-scoring match and then uses the
template string as the error, after substituting any group references.

There is a new page in the admin web UI for editing the list of rules,
and for testing new patterns against the last 1000 jobs.