Fix job cancellation
Before, we turned off the switch both when cancelling and when finishing
the job, but cancelling is only a request. In particular, we shouldn't
close the log or release pool resources just because the user has asked
us to cancel.

This commit removes the `switch` argument from the cache API. Instead,
use `Job.on_cancel` if you want to do something when the job is
cancelled. This is slightly clearer and also prevents the user from
turning off the switch themselves.

Add a separate `Job.cancel` operation to request cancellation of a job.