Add Input.map_result
Allows reporting errors in the right diagram box if you need to
post-process something. In a regular map we have to use a separate
box because the input box might also be used for something else.
But an Input hasn't been put into a box yet, so it must get its own.