Colour binds based on input state, not output state
Before, we considered a bind node to include everything contained inside
it, so if the final result was an error then the bind itself would have
an error state. This doesn't made much sense, since we actually return
the stage the bind evaluates to as the final result, and it's confusing
that some earlier pipeline box suddenly goes red for no obvious reason.

Also, it doesn't work with incremental evaluation at all, due to the
cycle: the operations inside the bind depend on the bind, but the bind
depends on the result of the operations.