Use a single pipeline in the unit tests
Instead of running the engine with a different pipeline for each test,
use a single pipeline that gets the current test's pipeline from a
variable. This avoids the need to cancel the watches explicitly (we
just change the pipeline to an empty one and let the engine remove them)
and makes it easier to support incremental evaluation in future.