Use Job.cancel directly
Running jobs already have a `cancel` operation, and it doesn't make
sense to cancel a non-running job, so just call that directly and remove
`cancel` from `actions`. This also means that it is now possible to
cancel a job that isn't needed by the pipeline (this might be useful
if a job isn't set to auto-cancel).