Evaluate individual test pipelines inside test_pipeline
This is needed for incremental evaluation so that the evaluation happens
during an update and the system knows that the test pipeline should only
exist during the test. Otherwise, it will see each test pipeline as a
separate top-level thing and try to keep all of them running.

This change is in a separate commit because a side-effect is that the
diagrams now show the pipeline-choosing step. This makes it easy to see
that the incremental changes don't affect the results.