Change the representation of terms to support incremental evaluation
Before, a term was a changeable value containing both the dynamic and
"static" analysis results. The static result included a unique ID that
changed whenever the result changed. This meant that any change had to
propagate to all later stages because the static analysis graph had
changed, even though the values might be the same.

In particular, this prevented optimising `list_map` by only recomputing
the values that changed, since every result needed to link back to the
latest version of the map input.

Now, a node has a fixed static part and a changeable dynamic part. Also,
the static part really is static now. Previously it contained anything
that should appear on the diagrams, including (for example) the current
colour of each box and the link to the current job.

Internally, we now also track where active and error reports come from.
This means that we can now show blocked actions on the diagrams without
duplicating the logic: an item is blocked if it's active or an error and
not the source of that state.

One result of this is that you no longer get a snapshot of the static
analysis. Instead, operations such as generating the dot graph or
calculating the state metrics are done on the live pipeline (which isn't
a problem).

Another change is that simplification of the analysis graph is now done
when the graph is generated. For example, we previously created graph
nodes for failed map operations but not for successful ones. Since the
graph is now static, we always create a node for maps, but skip them at
rendering time if the operation succeeded.

Although the analysis part of a node is now static, it can still contain
incremental parts. For example, `bind`'s static type contains an
incremental part with the pipeline it currently expands to. In fact,
bind is now represented as two nodes: one with the operation name that
goes before the expanded pipeline, and one afterwards (not shown) to
hold the result.