Show updating monitor jobs with a gradient background
A primitive can now report a status for the job separately from the
output status. This allows a job to check for updates in the background
while continuing to provide the old value as its output.

For now, this is only used by `Monitor` jobs (which have always updated
in the background, but previously didn't show any visual indication of
this activity in the diagrams).