Latch previous result for scheduled rebuilds
If a rebuild is triggered by the schedule, do the rebuild in the
background without changing the output to pending. This allows polling
for changes without disrupting the rest of the pipeline.

To support this, instead of invalidating the disk cache entry when
moving from the `Retry` state, we now invalidate it when moving into
the state (but only in cases where we don't want to latch the old
value).