List all dependencies in opam files
If we list a dependency in our dune file, ensure we have a corresponding
dependency in the opam file.

Patch generated by https://github.com/talex5/dune-opam-lint

I reordered the dependencies in a few places and updated version
constraints to match similar packages, e.g. cohttp and cohttp-lwt.
I also adding a missing `(package)` line to the tests, which was
adding a `fmt` test dependency on `current_incr` (which doesn't
matter at the moment, as it already depends on it via alcotest).