Provide Analysis.quick_stat for faster stats
`Analysis.stat` walks the graph and performs simplification so that the
counts match the visible boxes. However, this gets pretty slow with
large pipelines (about 3s for 10,000 boxes and 52s for 50,000 boxes on
my machine).

Instead, just keep counters for primitives and bind operations. This
makes getting the stats O(1). Use this for the Prometheus metrics.