Implement labelling of clusters on the Graphviz diagram.
Add an optional label argument to list_iter, list_map, option_map.

Pass this through the List_map and Option_map metadata options to the
label of the subgraph in the Graphviz input.

This changes the expected test results because the label is inherited
by subgraphs by default, so we need to explicitly wipe it with each
subgraph.

Signed-off-by: Ewan Mellor <ewan@tarides.com>