Explicitly handle and diagnose EPERM when killing child processes.
If the job uses sudo (e.g. a developer using "sudo docker") then the job
will not be cancellable.  In this case, we get EPERM from the kill call.
Diagnose this with a more explicit message.  Also, do not kill the
main process in this case, but simply let the thread exit.

Signed-off-by: Ewan Mellor <ewan@tarides.com>