Add a new look-and-feel for the web UI.
This adds Roboto / Roboto Mono as web fonts, and adds highlight colors
for table heading and the nav menu based on the OCaml logo.  It adds the
logo itself to the nav bar and the favicon.

For the Graphviz diagrams, set the node color (i.e. the box outline color)
to match the fill. This makes the boxes less heavy on the page, because
they no longer have the black border.  Set the default font for the
graph, node, and edge options to use Roboto.  This changes the unit
test expected output to include the options.

Lots of padding and spacing adjustments too.

Signed-off-by: Ewan Mellor <ewan@tarides.com>