Remove RO bit on Windows file when deleting tmpdir
OCurrent suffers from the same problem as Lwt used to, see Lwt PRs 849
and 861.

On Windows, when read-only files are created (Git marks its index
files as read-only), Lwt_unix.unlink will fail to delete these
files. This is annoying when cloning a Git repo in a call to
`Process.with_tmpdir`. The solution is to remove the RO bit by
chmod'ing the file.