Display line numbers in job logs
The job logs page is streamed from the server to the client, while the job is
running. This commit uses Mutation observers on the pre tag to append line
numbers to the newly added lines in the <pre> tag.

We also add IDs to the line number span tags, which allows jumping to specific
line using #L<num> in URL.