Fix some overuse of memory and concurrent fd open when too many jobs are queued
All the log files of queued and running jobs are let open and consumes a lot of memory while sitting idle, which causes problems for opam-repo-ci (which can have easily milions of jobs queued at times).

This fix makes it so that log files are reopened and closed everytime they are used instead of let open. This is a bit more syscall intensive but in practice opam-repo-ci has been relatively fast with this change for the past 6 months.