Expose reference (PR or MR) and associated HEAD commit for the CI
The plain `ci_ref` returns only the list of commits which are HEADs of
references (like the last commit of a PR). This lower-level function
exposes bindings between the reference and the commit for when the
pipeline needs both, e.g. when posting the build status of a commit to
the associated PR/MR.