Fetch most relevant PRs and branches by update dates
Currently, we fetch only the first 100 PRs and branches of a GitHub repository,
and there's a TODO in `plugins/github/api.ml` about using cursors to fix this.
It would be nice to have the TODO fixed, but meanwhile returning the most
relevant PRs and branches sorted by last update timestamp may be useful.  This
commit changes the API call to fetch the 100 most recently updated PRs and
branches, instead of the first 100 PRs and branches.